Make SPM integration tests remove the Packages folder
This will ensure the environmet is pristine every time.

I came to this becaues my local builds kept failing due to SPM not
finding any tag in the repo it fetched, despite the tag being there both
locally and on the remote.

This was due to SPM not pulling from the remote since it already had the
repo locally.

There probably is a flag that can be passed to `swift build` to make
sure that it pulls, but this works as well.